requirements.txt
flask
python-dotenv
requests
beautifulsoup4
transformers
torch
supabase
firecrawl-py
tldextract
